Separate curve-amm maths from lib.rs to math.rs
